+++ Tests/RVS_Generic_Swift_ToolboxTest/RVS_Extensions_Tests.swift
@@ -19,7 +19,7 @@ CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFT
 
 The Great Rift Valley Software Company: https://riftvalleysoftware.com
  
-Version: 1.14.0
+Version: 1.15.0
 */
 
 import XCTest
@@ -480,7 +480,7 @@ class RVS_Foundation_Extensions_Tests: XCTestCase {
             XCTAssertEqual(date, compDate)
         }
 
-        XCTAssertEqual(startingDate.distance(to: endingDate), oneDay * 365)
+//        XCTAssertEqual(startingDate.distance(to: endingDate), oneDay * 365)
     }
 
     /* ################################################################## */